v5/vibe-inc/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vibe-inc"
version = "0.1.0"
description = "Vibe Inc D2C marketing roles on OpenVibe V5"
requires-python = ">=3.13"
dependencies = [
    "openvibe-sdk>=1.0.0",
    "openvibe-runtime>=1.0.0",
    "facebook-business>=20.0.0",
    "google-analytics-data>=0.18.0",
    "ShopifyAPI>=12.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vibe_inc"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
